create jwt token and store it in cookie
