+++ api.py
@@ -4,7 +4,7 @@ from mysql.connector import pooling
 
 from API.attraction import attraction
 from API.category import category
-from API.user import user
+
 
 dbconfig = {
     "host" : "localhost",
@@ -24,4 +24,4 @@ api = Blueprint("api", __name__)
 
 api.register_blueprint(attraction, url_prefix = "")
 api.register_blueprint(category, url_prefix = "")
-api.register_blueprint(user, url_prefix = "")
+
