+++ static/js/user.js
@@ -1,4 +1,6 @@
-function register_btn(){
+// 註冊程序
+const register_btn = document.querySelector(".register_btn")
+register_btn.addEventListener("click",()=>{
     let register_info = {
         "name":register_form.name.value.toString(),
         "email":register_form.email.value.toString(),
@@ -15,16 +17,17 @@ function register_btn(){
     .then((response)=>{
         return response.json()
     }).then((data)=>{
-        console.log(data)
         if(data["ok"]){
             alert("註冊成功!")
             location.href = "/"
         }else{
-            alert("註冊失敗，請重新註冊。")
+            alert(data["message"])
         }
     })
-}
-function login_btn(){
+})
+// 登入程序
+const login_btn = document.querySelector(".login_btn")
+login_btn.addEventListener("click",()=>{
     let login_info = {
         "email": login_form.email.value.toString(),
         "password":login_form.password.value.toString()
@@ -39,15 +42,14 @@ function login_btn(){
     }).then(response=>{
         return response.json()
     }).then(data=>{
-        console.log(data)
-        console.log(document.cookie)
         if(data["ok"]){
             location.href = "/"
         }else{
-            alert("帳號密碼錯誤，請重新輸入")
+            alert(data["message"])
         }
+        
     })
-}
+})
 
 
 const grayscale_div = document.querySelector(".grayscale")
@@ -63,27 +65,31 @@ const switch_to_login = document.querySelector(".switch_to_login")
 reservation_btn.addEventListener("click", ()=>{
     location.href = "/booking"
 })
+// 顯示登入區塊
 signin.addEventListener("click", ()=>{
-    const nowHeight = document.querySelector("body").scrollHeight
+    let nowHeight = document.querySelector("body").scrollHeight
     grayscale_div.style.height = nowHeight + "px"
     grayscale_div.style.display = "block"
     login_background.style.display = "block"
 })
 login_background.addEventListener("click", ()=>{
-    const nowHeight = document.querySelector("body").scrollHeight
+    let nowHeight = document.querySelector("body").scrollHeight
     grayscale_div.style.height = nowHeight + "px"
     grayscale_div.style.display = "block"
     login_background.style.display = "block"
 })
+// 關閉按鈕
 close_btn.addEventListener("click", (e)=>{
     grayscale_div.style.display = "none"
     login_background.style.display = "none"
     e.stopPropagation()
 })
+// 切換註冊區塊
 switch_to_register.addEventListener("click",()=>{
     document.querySelector(".login_area").style.display = "none"
     document.querySelector(".register_area").style.display = "block"
 })
+// 切換登入區塊
 switch_to_login.addEventListener("click",()=>{
     document.querySelector(".login_area").style.display = "block"
     document.querySelector(".register_area").style.display = "none"
