+++ templates/index.html
@@ -35,7 +35,7 @@
                 <form id="login_form">
                     <input type="email" placeholder="輸入電子信箱" name="email"/>
                     <input type="password" placeholder="輸入密碼" name="password"/>
-                    <button type="button" onclick="login_btn()" class="login_btn">登入帳戶</button>
+                    <button type="button" class="login_btn">登入帳戶</button>
                 </form>
                 <a><div class="switch_to_register">還沒有帳戶？點此註冊</div></a>
             </div>
@@ -45,7 +45,7 @@
                     <input type="text" placeholder="輸入姓名" name="name"/>
                     <input type="email" placeholder="輸入電子信箱" name="email"/>
                     <input type="password" placeholder="輸入密碼" name="password"/>
-                    <button type="button" onclick="register_btn()" class="register_btn">註冊新帳戶</button>
+                    <button type="button" class="register_btn">註冊新帳戶</button>
                 </form>
                 <a><div class="switch_to_login">已經有帳戶了？點此登入</div></a>
             </div>
